+++ .storybook/preview.ts
@@ -13,6 +13,7 @@ import '../packages/ui/dist/css/components/link.css';
 import '../packages/ui/dist/css/components/text.css';
 import '../packages/ui/dist/css/components/body.css';
 import '../packages/ui/dist/css/components/eyebrow.css';
+import '../packages/ui/dist/css/components/caption.css';
 import '../packages/ui/dist/css/components/heading-display.css';
 import '../packages/ui/dist/css/components/heading-sans.css';
 import '../packages/ui/dist/css/components/heading-serif.css';